+++ example/project/settings.py
@@ -33,6 +33,7 @@ INSTALLED_APPS = [
     "counters",
     "coffee",
     "modals",
+    "registration",
     "uploads",
 ]
 
@@ -116,6 +117,7 @@ STATICFILES_DIRS = [
     BASE_DIR / "project/../counters/components",
     BASE_DIR / "project/../coffee/components",
     BASE_DIR / "project/../modals/components",
+    BASE_DIR / "project/../registration/components",
     BASE_DIR / "project/../uploads/components",
 ]
 STATIC_URL = "/static/"