+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## UNRELEASED
+
+- Added a "registration" app to show how to use livecomponents with Django forms.
+
 ## 1.11.1 (2024-02-26)
 
 - Added `livecomponents_login_required` decorator to require a user to be logged in to call a component method.